Feature/add the round file and move functions
