+++ src/card.js
@@ -1,5 +1,6 @@
 const { sampleData } = require("./sub-data");
 
+
 let createCard = (id, question, answers, correctAnswer) => {
   let card = {
     id:id, 
@@ -41,29 +42,24 @@ let createRound = data => {
 createRound(sampleData)
 
 
-const takeTurn = (guess, round) => {
-  let evaluate = evaluateGuess(guess, round.currentCard.correctAnswer);
-    if(evaluate === 'Incorrect!'){
-      round.incorrectGuesses.push(round.currentCard.id)
-   }
-    round.roundOver = true
-    round.turns += 1
-    round.percentage = calculatePercentCorrect(round)
-    round.currentCard = round.deck[round.turns]
-  return evaluate
-}
+// const takeTurn = (guess, round) => {
+//   let evaluate = evaluateGuess(guess, round.currentCard.correctAnswer);
+//     if(evaluate === 'Incorrect!'){
+//       round.incorrectGuesses.push(round.currentCard.id)
+//    }
+//     round.roundOver = true
+//     round.turns += 1
+//     round.percentage = calculatePercentCorrect(round)
+//     round.currentCard = round.deck[round.turns]
+//   return evaluate
+// }
 
 let calculatePercentCorrect = (round) => {
   round.percentage = (round.turns - round.incorrectGuesses.length)/round.turns * 100
  return round.percentage
 }
 
-let endRound = (round) =>{
-  if(round.roundOver){
-    let displayOver = `** Round over! ** You answered ${round.percentage}% of the questions correctly!`;
-    return displayOver
-  }
-}
+
 
 module.exports = {
   createCard, 
@@ -71,7 +67,6 @@ module.exports = {
   createDeck,
   countDeck, 
   createRound,
-  takeTurn,
-  calculatePercentCorrect, 
-  endRound
+  // takeTurn,
+  calculatePercentCorrect
 }
