+++ src/game.js
@@ -2,6 +2,10 @@ const data = require('./data');
 const prototypeQuestions = data.prototypeData;
 const util = require('./util');
 
+
+
+
+
 function printMessage(deck) {
   console.log(`Welcome to FlashCards! You are playing with ${countCards(deck)} cards.
   -----------------------------------------------------------------------`);
